Fix regex for excluding branches
With the formatting of `%(refname:short):%(upstream:track)` the output
will print all branches followed by a `:` and the tracking branch. To
properly ignore all `SKIP_BRANCHES` we want to add the `:` after any
matching value in the group and not just the last one.

This fixes a bug where the current regex is too greedy and matches the
beginning part only of all branches except the one specified last.